+++ src/react/useFormError.ts
@@ -1,6 +1,6 @@
 import { useStore } from "zustand/react";
 import { useShallow } from "zustand/react/shallow";
-import type { FormStateApi } from "./useFormState";
+import type { FormStateApi } from "./useFormSelector";
 
 export const useFormError = (
   form: FormStateApi,